fix performance degredation when paused
Paused frames were taking ~2.3 ms to draw,
which is 1 ms longer than when not paused.

I'm not sure of the root cause.
However the problem disappeared when I removed an if statement
that conditionally computes the states when not paused.
And now recomputing every frame when paused is faster
even though it's unnecesary.
Therefore I suspect it was due to some kind of branch misprediction,
but again I'm not sure.