add new balls on mouse click
I changed the inputs to the matrix exponential.
Instead of calculating xt from x0 and t,
we use x from the previous iteration and dt.
Mathematically this is the same, but I don't know enough
numerical analysis to tell if there are rounding issues.
Visually everything looks the same.

TODO: refactor