still working on the eigenvalue editor
There are some tricky parts, namely
synchronizing UI values and numerical computations
without running into rounding errors.
Complex eigenvalues are constrained to be in conjugate pairs
only if they exist.
This means we must be careful whenever the imaginary part is close to 0.

Currently the program will crash if we try to drag
either of the sliders for the imaginary components.