added matmul routine
We'll need this for the matrix exponential.

Reference:
The Scaling and Squaring Method for the Matrix Exponential Revisited
Nick Higham (2009)
DOI. 10.1137/090768539

I will first try a naive power series algorithm.
I started brainstorming some code for the the scaling & squaring algorithm,
but commented it out since it is more involved.