display an equation texture
It works on local machine but not itch.io.
There is an issue with cwd path in wasm code.
On itch.io the current working dir is /,
but it should be the location of the index.html file.

One possible way to get around this problem is
to embed the png pixel values as c source code,
and then convert it into a raylib texture.